Cleaning and standardizing the variable "Program."
Cleaning is based on using string distance calculations.

The cleaning involves considerable manual investigation of variables, and thereby DOES NOT CLEAN THE VARIABLE COMPLETELY. There are still many instances of misspellings.

At the end, an example of clustering is done to show how the clusters could be looked at instead of the variable entries themselves.